levels: address some issues in E1M5. (#1540)
It turns out all the secrets are marked using the thin door/wall sector instead of the part you stand in.
(The change in control sectors for the STARG secret is meaningless, I changed it and then realized what the clue was *supposed* to be and that this was actually really well done. Shame about the reward but I think that's beyond the scope of what I'm trying to do here.)

Addressed three obligatory points that almost require straferunning and precise mouse movement to reach in time:

1. The spaceship door switches are now set to stay open. The traps, etc. are governed by a bunch of trigger lines on all routes leading to it except the brown door.

2. The southeast torch lift is now operated by a switch right inside the platform. The original switch lowers a wall next to the lift so that the lift actually lowers when activated.

3. The gap jump between the two orange STARG walls is almost trivially easy with a mouse, a miserable savescum-fest without. Lowering the far one by 8 units makes it easy for everyone. (Not strictly obligatory since you can jump into the window from the south, Map20 style, but that's practically obscure enough to count as a secret.)

Fixed a misaligned texture on line 487.